requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pytz==2024.1
twilio==9.2.3
